requirements.txt
annotated-doc==0.0.4
click==8.3.2
markdown-it-py==4.0.0
mdurl==0.1.2
MouseInfo==0.1.3
mss==10.1.0
numpy==2.4.4
opencv-python==4.13.0.92
pillow==12.2.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.20.0
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
rich==14.3.3
shellingham==1.5.4
typer==0.24.1
